[feat] 플레이어 데미지 받는거 추가

+++ Assets/Scripts/TestFolder/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 {
-    public void TakeDamage(int Damage);
+    public void TakeDamage(float Damage);
 }
